refactor: move helpers to infrastructure
