+++ tests/console/app_container_test.py
@@ -2,7 +2,7 @@
 test global chk functions
 """
 from chk.infrastructure.app_container import App
-from chk.console.helper import data_set, data_get
+from chk.infrastructure.helper import data_set, data_get
 
 app = App()
